+++ develop/config/stylelint/template/ignore.content.js
@@ -1,3 +1,4 @@
+/* eslint-disable unicorn/prefer-module */
 /* eslint-disable import/no-commonjs */
 
 const content = ``;