+++ packages/antd-management-fast-component/src/components/MobileContainor/Devices/index.js
@@ -3,4 +3,5 @@ export * from './IPhone5S';
 export * from './Iphone8';
 export * from './Iphone8plus';
 export * from './IphoneX';
+export * from './NoneSketch';
 export * from './RoughSketch';