+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 2.6.84
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-common@2.6.55
+
+## 2.6.83
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-common@2.6.54
+
 ## 2.6.82
 
 ### Patch Changes